requirements.txt
streamlit
pillow
opencv-python-headless
numpy
weapon-detection-app/
│
├── app.py
├── model.py
├── requirements.txt
│
└── model/
    ├── frcn.weights
    ├── frcn_config.cfg